# Perf: Improved Performance of `ImagePainter`