This change improves performance by approximately 17%, but it uses about 11% more memory. The current implementation utilizes an Iterator, which is more memory-efficient, but slower.

```
Suites
+-----------+---------------------+--------+------------+--------+------------+------+-----------+----------+
| suite     | date                | php    | vcs branch | xdebug | iterations | revs | mode      | net_time |
+-----------+---------------------+--------+------------+--------+------------+------+-----------+----------+
| <current> | 2023-11-12 00:36:29 | 8.2.12 | image      | false  | 4          | 100  | 102.695ms | 10.297s  |
| main      | 2023-11-11 23:56:20 | 8.2.12 | main       | false  | 4          | 100  | 120.589ms | 12.004s  |
+-----------+---------------------+--------+------------+--------+------------+------+-----------+----------+
```

```
<current>
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+-----------+
| benchmark       | subject            | memory   | min       | max       | mode      | rstdev | stdev     |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+-----------+
| ImageShapeBench | benchImageShape () | 13.594mb | 102.455ms | 103.852ms | 102.695ms | ±0.51% | 528.971μs |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+-----------+

main
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
| benchmark       | subject            | memory   | min       | max       | mode      | rstdev | stdev   |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
| ImageShapeBench | benchImageShape () | 12.327mb | 117.841ms | 121.699ms | 120.589ms | ±1.17% | 1.404ms |
+-----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
``` 

So, this change is a pure trade-off between:

1) Memory vs Performance
2) Readability (the new implementation is slightly less readable)

So, it depends on the judgment for this part. Maybe a 17% improvement in performance isn't significant for this specif case (image rendering), and we might prefer to focus on using less memory and maintaining better readability, etc. :blush: 

My judgment: Since the changes are very simple, I tend to like this new version because of the performance improvement. Also, memory isn't a serious issue these days. But, on the other hand, I'm very confortable with the old implementation, using Iterators etc. =D