@dantleech 2023-11-12
I think for now it's better to leave it with the iterator -- the memory usage probably increases with the size of the image? what if you load a high res picture?

I think this problem probably goes away sufficiently when we don't have to re-render on each frame...