@dantleech 2023-11-12
confirmed, loading a larger image results in OOM where as it's ok with the iterator...