@KennedyTedesco 2023-11-12
@dantleech In theory, large, high-resolution images should be resized before their pixels are processed, right? The change in re-rendering will help a lot in one part, but it doesn't solve the problem of processing tens of thousands of pixels individually (as in the case of a large image).

But yes, sticking with the Iterator in this case is simpler, so there's no need to change anything for now.